requirements.txt
numpy>=1.16.2
lxml>=4.2.4
tensorflow>=1.13.1
tqdm>=4.31.1
Pillow>=6.0.0
